+++ dmucs/COSMIC/Sputs.c
@@ -12,7 +12,9 @@
  *               of this software.
  * Date:         Aug 22, 2005
  */
+#include <stdlib.h>
 #include <stdio.h>
+#include <string.h>
 #include "sockets.h"
 
 /* --------------------------------------------------------------------- */
@@ -32,7 +34,7 @@ Socket *skt;
 {
 
 /* write out buf and the null byte */
-Swrite(skt,buf,strlen(buf)+1);
+Swrite(skt,buf,(int)strlen(buf)+1);
 
 }
